# RazorSharedLib/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MtGApi",
    "version": "1.0"
  },
  "paths": {
    "/api/Log/get-top-50": {
      "get": {
        "tags": [
          "Log"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LogResponseRecordDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogResponseRecordDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogResponseRecordDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/MagicSet/save-set/{setCode}": {
      "post": {
        "tags": [
          "MagicSet"
        ],
        "parameters": [
          {
            "name": "setCode",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/MagicSet/get-set/{setCode}": {
      "get": {
        "tags": [
          "MagicSet"
        ],
        "parameters": [
          {
            "name": "setCode",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SingleMagicSetResponseRecordDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleMagicSetResponseRecordDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleMagicSetResponseRecordDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/MagicSet/get-set-list": {
      "get": {
        "tags": [
          "MagicSet"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MagicSetResponseRecordDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MagicSetResponseRecordDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MagicSetResponseRecordDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/MagicSet/Ping": {
      "get": {
        "tags": [
          "MagicSet"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PingResponseRecordDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PingResponseRecordDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PingResponseRecordDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Quiz/GetQuiz/{numOfCards}/{setCode}": {
      "get": {
        "tags": [
          "Quiz"
        ],
        "parameters": [
          {
            "name": "setCode",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "numOfCards",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MagicQuizDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MagicQuizDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MagicQuizDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "LogRecordDto": {
        "type": "object",
        "properties": {
          "createdUTC": {
            "type": "string",
            "format": "date-time"
          },
          "section": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LogResponseRecordDto": {
        "type": "object",
        "properties": {
          "logs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LogRecordDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MagicQuizCardDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "cmc": {
            "type": "integer",
            "format": "int32"
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "cardTypes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MagicQuizDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "setCode": {
            "type": "string",
            "nullable": true
          },
          "cards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MagicQuizCardDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MagicSetResponseRecordDto": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "sets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MtGSetRecordDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MtGCardRecordDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "rulings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MtGRulingRecord_DTO"
            },
            "nullable": true
          },
          "abilities": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "multiverseId": {
            "type": "string",
            "nullable": true
          },
          "subTypes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "types": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "superTypes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "cmc": {
            "type": "integer",
            "format": "int32"
          },
          "isColorLess": {
            "type": "boolean"
          },
          "isMultiColor": {
            "type": "boolean"
          },
          "manaCost": {
            "type": "string",
            "nullable": true
          },
          "setName": {
            "type": "string",
            "nullable": true
          },
          "set": {
            "type": "string",
            "nullable": true
          },
          "number": {
            "type": "string",
            "nullable": true
          },
          "legalities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MtGLegality"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MtGLegality": {
        "type": "object",
        "properties": {
          "format": {
            "type": "string",
            "nullable": true
          },
          "legalityName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MtGRulingRecord_DTO": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "nullable": true
          },
          "text": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MtGSetRecordDTO": {
        "type": "object",
        "properties": {
          "setName": {
            "type": "string",
            "nullable": true
          },
          "setCode": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PingResponseRecordDto": {
        "type": "object",
        "properties": {
          "answer": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "SingleMagicSetResponseRecordDto": {
        "type": "object",
        "properties": {
          "setName": {
            "type": "string",
            "nullable": true
          },
          "setCode": {
            "type": "string",
            "nullable": true
          },
          "cards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MtGCardRecordDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}